chore(web): extract shared logic of loadTemplate
