+++ .changeset/young-doodles-strive.md
@@ -1,0 +1,7 @@
+---
+"@lynx-js/web-core-server": patch
+"@lynx-js/web-constants": patch
+"@lynx-js/web-core": patch
+---
+
+chore: extract shared logic from web-core and web-core-server's loadTemplate into a unified generateTemplate function
