+++ packages/web-platform/web-constants/src/utils/index.ts
@@ -4,3 +4,4 @@
 
 export { LynxCrossThreadContext } from './LynxCrossThreadContext.js';
 export { dispatchMarkTiming, flushMarkTiming } from './markTiming.js';
+export { generateTemplate } from './generateTemplate.js';
